[FE-127] Refactoring all components.
