+++ src/components/Buttons/Buttons.jsx
@@ -4,13 +4,14 @@ import Deposit from './Deposit/Deposit';
 import Redeem from './Redeem/Redeem';
 import Repay from './Repay/Repay';
 
-function Buttons({ account }) {
+function Buttons({ account, onChange, userState }) {
+	debugger;
 	return (
 		<div>
-			<Deposit account={account} />
-			<Redeem account={account} />
-			<Borrow account={account} />
-			<Repay account={account} />
+			<Deposit account={account} onChange={onChange} userState={userState} />
+			<Redeem account={account} onChange={onChange} userState={userState} />
+			<Borrow account={account} onChange={onChange} userState={userState} />
+			<Repay account={account} onChange={onChange} userState={userState} />
 		</div>
 	);
 }
