+++ src/components/Buttons/Redeem/Redeem.jsx
@@ -3,12 +3,20 @@ import RedeemAll from './RedeemAll/RedeemAll';
 import RedeemUnderlyingAsset from './RedeemUnderlyingAsset/RedeemUnderlyingAsset';
 import RedeemWrappedToken from './RedeemWrappedToken/RedeemWrappedToken';
 
-function Redeem({ account }) {
+function Redeem({ account, onChange, userState }) {
 	return (
 		<div>
-			<RedeemAll account={account} />
-			<RedeemUnderlyingAsset account={account} />
-			<RedeemWrappedToken account={account} />
+			<RedeemAll account={account} onChange={onChange} userState={userState} />
+			<RedeemUnderlyingAsset
+				account={account}
+				onChange={onChange}
+				userState={userState}
+			/>
+			<RedeemWrappedToken
+				account={account}
+				onChange={onChange}
+				userState={userState}
+			/>
 		</div>
 	);
 }
