+++ src/components/Buttons/Repay/Repay.jsx
@@ -2,11 +2,15 @@ import React from 'react';
 import RepayAll from './RepayAll/RepayAll';
 import RepayUnderlyingAsset from './RepayUnderlying/RepayUnderlying';
 
-function Repay({ account }) {
+function Repay({ account, onChange, userState }) {
 	return (
 		<div>
-			<RepayAll account={account} />
-			<RepayUnderlyingAsset account={account} />
+			<RepayAll account={account} onChange={onChange} userState={userState} />
+			<RepayUnderlyingAsset
+				account={account}
+				onChange={onChange}
+				userState={userState}
+			/>
 		</div>
 	);
 }
