+++ src/components/ContentUser/ContentUser.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Table, Grid } from 'semantic-ui-react';
 import { SUPPORT_CURRENCIES } from '../../util/constants';
 
-import BU from './BalanceUser/BU';
-import BBU from './BalanceBorrowUser/BBU';
-import Col from './Collateral/Col';
+import BalanceUser from './BalanceUser/BalanceUser';
+import BalanceBorrowUser from './BalanceBorrowUser/BalanceBorrowUser';
+import Collateral from './Collateral/Collateral';
 
 function ContentUser({ account }) {
 	return (
@@ -29,13 +29,25 @@ function ContentUser({ account }) {
 							<Table.Row key={index + 123}>
 								<Table.Cell key={index}>{asset}</Table.Cell>
 								<Table.Cell key={index + 100}>
-									<BU key={account + index} account={account} asset={asset} />
+									<BalanceUser
+										key={account + index}
+										account={account}
+										asset={asset}
+									/>
 								</Table.Cell>
 								<Table.Cell key={index + 1000}>
-									<BBU key={account + index} account={account} asset={asset} />
+									<BalanceBorrowUser
+										key={account + index}
+										account={account}
+										asset={asset}
+									/>
 								</Table.Cell>
 								<Table.Cell key={index + 10000}>
-									<Col key={account + index} account={account} asset={asset} />
+									<Collateral
+										key={account + index}
+										account={account}
+										asset={asset}
+									/>
 								</Table.Cell>
 							</Table.Row>
 						))}
